refector: 열거형 클래스 이름 변경
기존의 enum LottoResult가 LottoRank로 변경되었습니다.
